Revert "Floss: implement le_audio_shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30b6bb85-5ead-4530-ab29-f53bc439f56a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339380984
Change-Id: I10fa0a89b637b24924d0f073a7e4720a5e857ec7