Revert "Floss: implement le_audio_shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1cf49e33-ab91-4159-81cf-3660cf36bfb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339385103
Change-Id: I056c708759fba3e0a7eac7e97d93a5b31637cab7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com