Revert "flags: Add device_iot_config_logging flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5639caa7-2fde-49f1-a5b4-ab96d4f57020). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id2780ed2b0cdb27ff2ec297bfe0fab52891effc7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com