Revert "Floss: implement le_audio_shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb35611b-5c2c-406b-af12-651d4dc0dbf0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339384724
Change-Id: I4a9b1375e7ffa5596b1a41f597f646c03eaebca8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com