AdapterService: Load native when ready to use
We want to delay loading the native to when we actually enable the
bluetooth. This remove operation from the onBind and decrease the risk
of binding failure.

Note that a failure during binding is never catch by the manager service
as there is no linkToDeath yet

Bug: 311772251
Test: atest pts-bot,avatar | Lifecycle is intensively tested there
Flag: Exempt, heavily test lifecycle change
Change-Id: I3636bae16a8854fad50bfbb920b1dd12fbf8fb11