Revert "floss: Complete libflags shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be2707f1-7daf-4cdf-9af0-2b5d1457bd24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339739862
Change-Id: Iff015b9e8dd9b9a2ce1f07b3df2346aee2ebb070
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com