Revert "floss: Complete libflags shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff86fa0c-1505-4f37-821d-749c94dc7fd4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339739525
Change-Id: I49866c415619efe89d84e2bc798fde25a059808c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com