Revert "floss: Complete libflags shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a4d2b87-43bc-42af-b86d-ea4bcb8d0d05). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339739805
Change-Id: I5d99ac07c14e01b77463dd5caf28a5970fa4d659
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com