Revert "floss: Complete libflags shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48710bfb-9af8-46fc-b048-a9eee72b0680). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339739827
Change-Id: I96c15f94a21c4c8b6e73466498740e3906284236
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com