Revert "floss: Complete libflags shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/547f5784-a741-455a-8376-5fcc4ad8598b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339739806
Change-Id: If3661c344e0ebc3650852b6d841f585e1f067df6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com