Revert "floss: Complete libflags shim"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d14e762-48eb-46ec-99ca-6153fdcbb811). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 339744828
Change-Id: I26725c74c2f42bf230fae430e47f747bd2254101
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com