Fix BluetoothOppTransferTest
Change oppConnectionReceiver_onReceiveWithActionSdpRecord_* test to
verify when device address is not null and no sdp record is passed to
intent. The test is renamed to _withoutSdpRecord accordingly, and is
changed to only run when flag identity_address_null_if_unknown is
enabled, since that is needed to ensure that device address is not null.

Bug: 295907764
Flags: EXEMPT, test fix only
Test: atest BluetoothOppTransferTest
Change-Id: I0bc6bdc4a05ceadab5b26c9f1a310d69d1cd4d1a