Revert "Add sdp_finished helper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1fd4948-e209-4a6a-a08f-083a7799a14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340565552
Change-Id: I4c7c3f596c657ec07063e07787bd04d051be4aa0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com