Revert "stack::sdp [5/12] handle.SDP_AddAttribute"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/547b44ba-0b71-4cdd-b008-23c9770fd46b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340685838
Change-Id: I538ce784efcbfba478c331a0d66dc6245f3a6b2f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com