hh: run all uhid exchanges in uhid thread
Introduce a new fd to be used for internal events between btif and
uhid thread. Use this new communication channel to serialize uhid
requests and response so all uhid related variables are only accessed
in the uhid thread.

NOT TESTED - DON'T REVIEW

Probably I made some silly mistake somewhere and I'd be super surprised
if this works on the get go.

Bug: 291522341
Test: m -j

Change-Id: I38073d4ad064c9b30103a070efdad87353a8c9f5