Revert "stack::sdp [5/12] handle.SDP_AddAttribute"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b29ca5c-9bf9-4234-8902-3c23a9b135a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia68850e6cb356a24f29fb1c55e1f8e7618e472d4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com