Floss: Always apply the +CIEV workaround, unless MPS mode is enabled.
To avoid conflicts between the telephony feature and the +CIEV
workaround, the +CIEV workaround was previously removed when the telephony
flag was enabled. However, since the telephony feature is going to be
enabled by default, this could potentially affect users.

This change applies the +CIEV workaround regardless of whether the
telephony flag is enabled. This may cause some telephony features to
malfunction, but these issues will be fixed in subsequent patches.
   
Bug: 340375662
Test: Conduct the following manual tests
 - enable bluetooth telephony
 - pair Bluetooth Headset need +CIEV command to work(Parrot Zik 2.0)
 - playback with youtube and online voice recorder and make sure hfp
 audio work
 - enable uhid and observe the hfp audio will not work
Test: atest bluetooth_test_gd
Tag: #floss
Flag: EXEMPT, no behavior change on Android; Floss-only changes

Change-Id: I1e97c5a8ee6d9465f68bd505132e7c2d1b2836ed