Revert "Add sdp_finished helper"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f0cba619-4228-45c7-baa8-c4fc83a67c13). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 340820079
Change-Id: I7fdf79653beae4445059bc2ed2720ec752905763
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com