Revert "RemoteDevices: Add logging to fetchUuidsWithSdp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/579477db-c653-424a-9152-4ecf10fc5c11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I52488811e873e454de08ef6f9d7878a1579d9850
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com