BassClientService: Remove old broadcaster when new synced on the same syncHandle
In case when onSyncEstablished (new broadcaster) came before onSyncLost (old
broadcaster) on the same syncHandle, properly handle retrieve broadcast ID and
remove the old one with notification.

When onSyncLost came, the new broadcaster will be removed but on next
onScanResult will be added again. On the controller nothing change as on
native it is still synced. Such operation is seamless for the user.

Bug: 342552854
Bug: 333691526
Test: atest BassClientServiceTest
Change-Id: I222f35b813132840871088e4ef15fa0cd2f9d15b