Revert "stack::l2cap Enforce API checks with nodiscard"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27309b03-9920-49bd-b654-e5c9bc1b29a8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4cac5dcbc0f253c1d830713716a14454070ad523
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com