Revert "AdapterServiceTest: Handle cleanup at shutdown"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c43b38b7-08b1-47ef-88f0-0c6c1a1fe3fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343561006
Change-Id: I747271b22afb85b2d866da254e591bca529198a7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com