system: Introduce log::assert_that_softly
assert_that_softly provides flexibility in handling
assertion levels, allowing different backends to interpret
the severity of assertion differently. This addresses the
the varying expectation in Bluetooth controllers behavior.

For Android, assert_that_softly behaves the same as assert_that.
For Floss, assert_that_softly is a relaxed version that shows
an error log if the condtion is not met, but doesn't abort.

Test: m com.android.btservices
Bug: 305066880
Flag: EXEMPT, mechanical refactor

Change-Id: I45a1e1d4762e1292ea0daa4903bc0585d6fbfa46