Revert "AdapterServiceTest: Handle cleanup at shutdown"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8ee9cc8-84c2-43b9-9703-be70cbd6760c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343559188
Change-Id: Ic32eae741bdc92e86154d8f5b0c753ecdb5a3b5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com