Revert "Floss: fix MSBC temp packet decode buffer size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4fbae95-bb16-45d2-93e0-bc0c2e1d5a1b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 344074770
Change-Id: I2f17c2dd9d2260a566b7c90ca1cca4448618656d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com