Revert "system::stack: Comment out unused parameters"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e708f604-9990-4a02-a924-a6826e03a207). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6aa8abb5a6963ba3bfcf9772defd8765f636cddb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com