Revert "system::stack: Comment out unused parameters"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9674fcb2-8593-4679-9079-1849fe3f9c8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33189c8e7087eefd2f9a2d510ccba9e43f15d9dc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com