Revert "system/btif/src: Update FAIL with more descriptive PARM_..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9674fcb2-8593-4679-9079-1849fe3f9c8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I06f9c23165c4f4fbed1e41dac79fda57caddf3ac
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com