Revert "system::stack: Comment out unused parameters"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e15c46d7-2c8d-4790-bd76-11e3f0ecb974). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf44d6b9db94078ecdb3e721a1fe9c63c6ba7702
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com