Revert "system/btif/src: Update FAIL with more descriptive PARM_..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e15c46d7-2c8d-4790-bd76-11e3f0ecb974). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a8cc51262143a21b2bdc68f0dfbfb3d44c0f239
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com