Revert "IWYU: Remove base::StringPrintf from tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/405be544-c7bf-44a8-b99d-c1d872936853). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4d3f32a8367051dd45bf2e7d02f39d445c97f50
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com