Revert "floss: topshim: Add utility for easier sharing types wit..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cfb7eb28-834c-4a7c-b0e4-f8af7225b907). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I52e5fd5216f84c5ae64cb4ca275518ab7e2a605a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com