Revert "floss: topshim: gatt: Remove unnecessary type convert"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cfb7eb28-834c-4a7c-b0e4-f8af7225b907). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I70ca647bd63099a1365c7663c6872e03854214ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com