Revert "floss: topshim: Add utility for easier sharing types wit..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cfb7eb28-834c-4a7c-b0e4-f8af7225b907). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I47452101023ab8c3b17c0ce2ab50ed61b5c9a1fb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com