Revert "system::stack: Comment out unused parameters"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ae92dbd-b8e3-4730-a8ae-97664ba88793). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic34871cf5fc431f59b40ce657ce8b8c355e85919
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com