Revert "system/btif/src: Update FAIL with more descriptive PARM_..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ae92dbd-b8e3-4730-a8ae-97664ba88793). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62fec16829e5430379bfc2f477d0d862690b6310
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com