Floss: Wait to request battery data until after bonding
Previously, we would send requests to read GATT-based battery data
immediately after forming an ACL connection. However, if the device was
not bonded by the time this request was sent, some peripherals would
reject the read request.

With this change, we wait until after bonding is complete to request
battery data.

Bug: 329796289
Test: m && manual verification with Logitech Casa Keys
Change-Id: Iaf29059c1b3149cf54d9d6a22ed5988b3d339d87