ScanMode: Stop writing ScanMode in config
The scan mode is erased when the stack shutdown, and it is force
overwrite when the stack start. It serve no purpose to write it on disk.
Prior to this, turning off the stack is generating a config change and
it is slowing down the shutdown, especially on slow disk hardware

Bug: 310147428
Test: none
Flag: Exempt refactor
Change-Id: I405d973f8ce9bbfbd35fff0a4aab2da8bcf54203