hh: Queue input data when before receiving UHID_OPEN
UHID_OPEN is sent by UHID when it's ready to receive data. It is
possible for a peer device to start transferring data even before we
receive UHID_OPEN. In such a case, the input will be lost.

Introduce a queue to temporarily store the input data when UHID is not
ready. Additionally, add a small delay even after receiving UHID_OPEN
because it is known for some cases that the underlying input could
still be not ready even after UHID_OPEN is received.

Bug: 299192582
Bug: 291522341
Bug: 345498287
Test: mmm packages/modules/Bluetooth
Test: see test in aosp/3003554

Change-Id: I136c7800e3290dd5aa521eea9fa57a7f5d814aa2