Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a6575b2-e582-4694-b3a3-cbc66a180df2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f29a70381e3ea8e3792458216c18913f3b7d9ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com