Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fc388080-9fd3-4ca7-8378-1803d156aad6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I50d26e2ba8ef67ade974d22ad6b57ac86be30b57
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com