Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ecedc251-1dac-432a-a116-a800f6128a12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7373fc4746cbb3b41c58c487c839b114dc62828b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com