Revert "MapClientStateMachineTest: Send correct message in testR..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1b245fd-e34d-434c-8869-6a563b6b19ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id64c667005939666d915ff1c57645202a4a063cc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com