Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/06c3073f-04d3-4112-b62e-8af587847afc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a686a7d259451677d2b5c08b7e0ff57b9d29ef5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com