Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/085d7910-b96f-4166-93aa-b759f9f85c37). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I371c05b4ad90bee7e851a6351331418daf4f3ef1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com