Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e02b85bc-7510-4629-8d74-ee347e8c8db1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347862795
Change-Id: Icdcd9b3754835e0137b21c07e6f39bbd12511acc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com