Revert "Get rid of delay_bonding_when_busy flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc4d474d-b9a4-424f-af62-61ee11e6e9d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id476338e0f5790608004ee3349d7ada8adde19a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com