Revert "Get rid of delay_bonding_when_busy flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc4d474d-b9a4-424f-af62-61ee11e6e9d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ba732a19d58d21e0ac52efb11cf9ea356242182
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com