Revert "Get rid of le_scan_parameters_fix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc4d474d-b9a4-424f-af62-61ee11e6e9d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9e6b9a74686e6e80a20e8e0290483a5f40a2c48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com