Revert "Get rid of le_audio_fast_bond_params"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc4d474d-b9a4-424f-af62-61ee11e6e9d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iecc03be5f2d275433d0179e04bd80a105129a65b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com