Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84367bda-ba49-40ba-9c03-c32602d56072). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If915cf134425f35f53b061d6f5d83448d73c1247
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com