LeAudio: Fix crash on stack shutdown
The unicast profile always stops the CodecManager first, and
if the broadcaster tries to update the hal client instance in
the codec manager the container object holding that pointer
is already destroyed. This is fine as the stack is shutting down
anyway and we should not assert because of that in the broadcas
service cleanup.

Bug: 338937189
Test: atest bluetooth_le_audio_client_test
Flag: Exempt; trivial fix
Change-Id: I893694717035983fb079377752fb1e9f89e9b8d7