Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efa6aa3d-c2f9-4da3-983b-8a8e11c79028). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief65b523c85e34a0c6f13d52979a3a91ecb1c76d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com