le_audio: LeAudioService: Fix handling initial zero available contexts
This fixes handling initial zero contexts that may be reported once
device is paired. Then the contexts are back, the Phone would like to
activate the device. Thus we'd like to handle this device just like
the onces that the stack inactivate due to available contexts that
are gone.

Bug: 357472821
Flag: com.android.bluetooth.flags.leaudio_unicast_no_available_contexts
Test: atest com.android.bluetooth.le_audio.LeAudioServiceTest
Change-Id: I62f913573dd6a8ba2c3bbbe3a578f7a2d4ecb0dc