le_audio: Fix trying to remove CIG when there is still CIS connected
This fixes ClearGroup method that called LE Remove CIG command
on CIG that had CISes still connected.

As per Core v5.4 | Vol 4, Part E; 7.8.100:
"If the Host tries to remove a CIG which is in the active state, then the Controller
shall return the error code Command Disallowed (0x0C)."

As the command failed, the CIG was not removed and left in REMOVING state.

Bug: 357472821
Flag: TBD
Test: atest bluetooth_le_audio_test
Change-Id: I3a36aa669ecff87741a43dfb1038c60b62d73b1c