leaudio: Remove autonomous disable timeout
Autonomous disable timeout turns out to be not needed.
The way it was implemented, would block remote device to disable one
direction for the bi-directional CISes e.g. when remote decide to no use
microphone.

Also the way it was implemented did not help with invalid remote behavior
which would be not moving to QoS Configured state when CIS was lost.
In fact it is not needed, as Android assumes that after CIS link lost,
ASEs are in the QoS Configured state as per specification. Remote
notification on this state does not matter here - which helps in IOP.

Bug: 335799807
Bug: 343860840
Flag: Exempt, regression tested with unittests
Test: atest bluetooth_le_audio_test

Change-Id: I01091291398b8cc2c08749c6d0533079672d2597