btif_av: Make public the methods from btav_sink_interface_t
The interfaces btif_av_get_snk_interface() and
get_profile_interface(("a2dp_sink") are not used
except in the JNI layer.

The call to get_profile_interface in com_android_bluetooth_a2dp_sink
is removed in order to de-obfuscate the call graph for a2dp.

The inferface type and declaration is also removed now that it is
unused.

Bug: 331817295
Test: m com.android.btservices
Flag: EXEMPT, no logical change
Change-Id: I6d065f3ee0eaee3c521d297925c2f504159d8a22