Revert "RootCanal: Reset APCF scanner in LinkLayerController::Re..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d1133a92-c95d-4db4-be70-e2f5a1b1c855). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb04452a2795149bbe12c4c89844861d7207772f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com