Revert "RootCanal: Reset APCF scanner in LinkLayerController::Re..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2439cb43-d458-45a9-9367-8e043a12adba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f1a66dc8e721fd064ebc186cb3a56c3fa5c3f4c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com