Revert "Add support for SuspendStream"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25f8fb94-8300-4b24-80e0-524595f03720). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie33381b30e7886b975c7eb604b94676d0252784f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com