Revert "Add support for SuspendStream"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1171460a-e6b9-461f-8381-a3d9cadb6b54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I522a2abc9c0b0a988b3579a730e4dd383faa0d08
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com