Revert "Add support for SuspendStream"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c20e2091-8d36-4d11-964b-07e53e084676). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia25cfe0dad328d36a9d4c4efc55965da6b4252c5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com