Revert "Floss: generalize LC3 decode buffer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64e82ddf-d72f-4192-9024-50a8588b6d77). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24f503067f0063cfa61d79e0a771578960439490
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com