Remove device only upon the last cancellation
While multiple clients attempt to connect to the
same device, if one client cancels, the device
should not be removed from the accept list,
as this would prevent other clients from connecting.

This change checks if other clients are still
interested in the device and avoids removing it
from the accept list in such cases.

Bug: 349232327
Bug: 349234039
Test: atest GattClientTest
Change-Id: If5fea38d6034b581c7a401351090c65305af3476