Change inband ringtone based with active device
Previously when there is no LE Audio, if there is 1 connected HFP device, it will always be active. But with LE Audio, it might be connected but not active.
In such case, phone should notify the remote that in band ringtone will not be played. It should use out band ringtone.

Bug: 346498545
Bug: 349318724
Test: atest HeadsetServiceAndStateMachineTest
Flag: com.android.bluetooth.flags.ag_provide_in_band_ringtone
Change-Id: I1470bf06b0639732eab48447b134694123241f58