Revert "start HFP offload session when active device becomes not..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e58986cd-b3c8-46a0-af75-83d3fa467ad7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c6c208356226abb34fc49194ee24a9510100b55
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com