Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/35d9713e-1567-4e97-86f0-9ef535b5f477). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347861095
Change-Id: Ia16a78c8d4543b78b0894c33aee38cd882798b32
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com