Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02c3a387-6c98-4cfb-9cff-6deb8e6f1047). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347861095
Change-Id: I7fa0d23419fde4f18dcac626fd2596a82937b7e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com