Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a525f0d-9ebd-4753-83c5-c2b7b67c0853). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347836801
Change-Id: Ib14e74b6fd7e95059db1a84321caf00a2cbd345e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com