Revert "Adapter: move register callback to oneway call"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc3d09ef-4611-4d19-b42f-8396d8a443d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I369dd2ac85684c6090f1bb939aa94365275a0a9c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com