Revert "SystemServer: SatelliteMode bypass thread jump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8759c62e-9914-4874-b656-e826e14dcf26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349684439
Change-Id: I6b671c3a626344e61378caedf01ac2bc8a805c69
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com