Revert "add kSamsungDebugHandle to filter debug logs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1136c699-6ab2-4ac0-94e3-8b9eabb8d888). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3225ee38d43b63d5982c3379999798eea6133569
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com