Split sec_state into link-specific variables
Currently when both LE and Classic links are established at same time,
and both are encrypted at same time, we have a problem. Same variable is
used to mark state as "encrypting" for both links, and it can cause race
conditions.

To fix that, keep security state as separate variable for LE and Classic
link.

This patch also removes security states that were added to address
similar issues around disconnection.

Bug: 343639888
Test: establish LE and Classic connection to dual mode device, verify
      encryption and profiles are connected on both transports
Test: m -j32
Change-Id: I0de9075278010f36a332156a8d984efba08fac00