Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21d81d61-e7c2-47f3-8d47-c6166d0dda78). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90ec7bd0e9522dbed56bdfceb7462aebdc47857d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com