Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/65cbcf83-65c2-420f-ac9a-5c44a91bc515). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab64c422cb85b6eb48c208abe78dfc2f412aa38a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com