Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c832815-1bd3-4368-b609-227bc46ea591). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I956beec0cbb331b5a7cd340d7d97fa4b6596d6ed
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com