bta: le_audio: state_machine: Ignore unexpected CIS established event
This fixes unexpected CIS established event handling that may be received
if already scheduled by the controller, even if the Host issued
CIS disconnection before.
This may happen if the CIS has been already created, but not yet
reported to the Host.
In such corner case it's sane to continue streaming.

Bug: TBD
Flag: TBD
Test: atest bluetooth_le_audio_test
Change-Id: I94a864a933d6ec711c5d4e8095412300e3247f75