Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c5b68fe7-1dde-46b6-abb3-98c88e1f6184). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic314914a6e97e2925506150104f3581e88777135
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com