Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba25c5c8-aa9f-49d7-a9ac-72491a98238e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d09ae7896fc7b8b14fdb75e2cc8490a01dae32c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com