Revert "stack::btm get_btm_client_interface().sco.BTM_WriteVoice..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/162380c7-91dd-4f71-ada4-39446a068e16). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00787a799f4454949edd69a845ad4662c6a5d53d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com