Revert "stack::btm get_btm_client_interface().sco.BTM_WriteVoice..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb9b9d4b-8100-4bda-8172-193d4856d0a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4204423ab8f0d6b19ded11a9e1d64a24513a8b76
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com