Revert "le_audio: lowering the value of ISO suspend time"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb9b9d4b-8100-4bda-8172-193d4856d0a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic99ec293f27b283411ed035f185eae546f3f713b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com