Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1466059d-16a7-418d-bd5c-e80c4c8f150d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I930042ba2e1bf84a5bf398ca2f30ce1066ac3e3d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com