Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e39f5d55-c68b-4ea0-9c33-3636b4f16ed1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ed90857078633cea65c6f81b2b6b6674e1b702c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com