Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff286f31-e23b-4b1d-bb5d-c6062616d14b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I675fb9125d1086ed1747847bc93021100c328e2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com