Revert "stack::btm get_btm_client_interface().sco.BTM_WriteVoice..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aacb8509-a03f-4521-af1f-7c10250bece2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d4efee464914ce2337c60ff232840547920d440
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com