Revert "le_audio: lowering the value of ISO suspend time"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aacb8509-a03f-4521-af1f-7c10250bece2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifb13e00806c4b3d9c6c27012317945a7465d0616
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com