Revert "Refactor the scanner map to reduce gatt/scan dependencies."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4781483f-6f98-4be9-ba41-3a8ee958935c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iadd86e5e87645854a1170d8f3ebbdbbc0e7aa2d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com