Revert "Ensures handling messages"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b89af7c6-28c8-4c18-837b-244fa9106a59). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a9a2bf1ffb7968159d413d7279ebfd5896f8953
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com