GATT Read Multiple Variable Request simulation
Devices that support EATT are mandated to support "Read Multiple
Variable Length Characteristic Values".
For devices that don't support EATT, assume they don't support it for
now.

More fine grained approach is possible - we can send request, and if
remote reply with "Request Not Supported" error code do regular reads.
This would however require storing this information about device. Migt
be considered as solution in future.

Bug: 296836982
Flag: exempt, this is not used yet, will be under flag in code that
      makes this call
Test: manual, use LE Audio with device incapable of EATT
Change-Id: I5b170d0f12f71ca04bccc98c0037f6f79e0d8fa7