le_audio: Avoid sound effects while adding/switching broadcast source
When listening to the external broadcast, we should not play sound
effects for handover. This is a follow-up improvement change with aosp/2985512 to cover more corner cases.

We should block sound effect once started to add source or perform
source switching, otherwise
we might facing some race condition that we're resuming CIS for sound
effects and buds are in transition for external broadcast.

If adding source failed or source got removed, the original logic will
restore the context mask.

Tag: #bug
Bug: 351095373
Bug: 336468573
Test: atest BassClientServiceTest BassClientStateMachineTest
Test: manually verified by QA
Change-Id: Ibe1726515be51b261ea6985559a374103d29b5ac