Wait for cleanup to finish for A2DP source profile init
Problem: Due to recent change aosp/3141565, the cleanup method is
called in addition to init calls. The cleanup method call runs in
a separate thread and could result for it to be invoked post init
This results in source profile to be torn down and be in disabled
state.

Solution: Wait for cleanup to be done before invoking init method.

Bug: 352106509
Test:  m (lunch seahawk-trunk-userdebug)
FLAG_EXEMPT: Hotfix for aosp/3141565.

Change-Id: I49dc01c151965f8b75baf41a585faabc0763e938