has_client.cc: do not allow to select not available preset
Accoring to HAP v1.0, section 5.5.4:
"The HARC shall not set the Index parameter to the Index of a preset
record that has the isAvailable field set to 0b0 (the preset is unavailable),
or to an Index value that is not supported by the HA."

Calling `selectPreset` on any index resulted in executing Write to
Control Point characteristic. Now, procedure is not executed when
given index doesn't exist for specified device or `isAvailable` flag is
not set.

Bug: b/350880296
Flag: TEST_ONLY
Test: atest pts-bot:HAP/HARC/PRE/BV-17-C
Change-Id: I89c6aa9fa11c1d9ffabb522a5c521d3c316a8b65