Revert "Add flag rfcomm_always_disc_initiator_in_disc_wait_ua"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41e76019-3188-4718-89f1-3c1290072d39). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6db3158b4d676065e030b98d1815ad32b24c6a2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com