Revert "Don't abort pairing due to unrelated name discovery"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41e76019-3188-4718-89f1-3c1290072d39). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f36316371ca3b4e6ed66c98dace8c5b45a3f5b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com