Revert "Remove checks for 16-bit UUID in HOGP host"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41e76019-3188-4718-89f1-3c1290072d39). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I30627fe906295647d2794db7f00447780928dde9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com