Revert "a2dp_vendor_aptx_linux: Add functions that are accidenta..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c317f0d9-47fb-4097-9240-de92fdf4c22a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3094914a8e48de611cf613f4ee984396126d6c9a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com