vc: Fix handling volume setup after reconnect
Before the patch, after reconnection the Android always tried to set the
volume which is present on the earbuds dispite the Persistent Flag.
However, there was a use case, where value was not set properly  i.e
when after reconnection, Phone had aleady a valid cache valome.

To fix it properly there is a need to provide Persistent Flag to the
Java layer, so Android can make a proper decision.

With this patch we make sure volume of the remote device always will be
applied to the system after reconnection unless Volume Persist Flag is
set to Reset (0x00) and volume on the remote device is 0.

Bug: 352647510
Flag: Exempt, corner case, regressions verified with unit tests, new test added
Test: atest bluetooth_vc_test VolumeControlServiceTest
VolumeControlNativeTest
Change-Id: Ibdc82e3a6c7b7b330c066dce7328ed7ba38d0576

Change-Id: I30d36e188574a7ab5c12ee19a0587030fd793c25