bta: le_audio: Add Dynamic Coordinated Sets feature implementation
This adds the implementation of Dynamic Coordinated Sets support in client.cc.
This includes functionality of audio switching between devices in the set.
Typical scenario is device streaming when becoming inactivem sends PACS
Available Contexts notification with the contexts cleared and CSIS Dynamic
Set Properties Notification with active flag cleared. The ASE can be later
released by peripheral device, so that Phone can disconnect the CIS.
Once the CIS is disconnected and ASE released, the Phone looks for
anothe device in the set that can substitute the device that become
non-available for stream.
The substitute if any is attached to the stream if still active.

Change-Id: Id93f77a796ce29072b11790aa42801af1f8f4f33