hh: Introduce a delay property before acknowledging UHID_OPEN
It is observed that UHID_OPEN doesn't always mean the input driver is
actually ready to accept data from the HID peripheral. Therefore, if
we just forward the input after receiving UHID_OPEN, there is still
some possibility the input won't be processed.

Because it's difficult to know exactly when the input driver is
actually ready, and there's a requirement to not miss any input for
several products, one simple approach would be to wait for a short
time after receiving the UHID_OPEN event while queuing the incoming
input, and flush the queue once a certain timeout is reached.

This CL introduces a customizable delay properly for that purpose.

Flag: com.android.bluetooth.flags.hid_report_queuing
Bug: 299192582
Bug: 345498287
Test: mmm packages/modules/Bluetooth
Test: see test in aosp/3003554

Change-Id: Iee77fc8d2e3ffe1d3f042ecd3a55710e81695450