bta: le_audio: Do not stop streaming when CIS create is cancelled
This fixes an issue that caused audio stream to be stopped because
stack got CIS Established event with non-zero - Operation Cancelled
by Local Host error code.
It may happen that during attacment to stream procedure, ASE being
in ENABLING state, the ASE is Released by peripheral. If that happens
the Phone can cancel CIS establishment, so when the CIS Established
will carry the error, it has to be handled, so that audio keeps
streaming.

Change-Id: I268671359b21bd1abe0c591568710b91fd344955