Revert "[7/7] stack::btm::inq BTM_BR_INQUIRY_MASK => BTM_GENERAL..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/979787f3-d642-4acd-b788-23d462f83b40). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ee32dec9ad61cec80d68c398f3ea49726691319
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com