Send the correct mask for service data to the native layer.
The UUID portion of the service data mask should have all bits to 1, to ensure that the provided UUID always must match exactly. Note that the UUID matching exactly is already enforced on the ScanResult in ScanFilter#matches. But passing down the correct mask could allow for further optimizations in the lower layers.

Bug: 348818819
Bug: 346694914
Flag: EXEMPT, obvious bug fix
Test: atest BluetoothInstrumentationTests, atest BumbleBluetoothTests, atest pts-bot:GATT
Change-Id: I1ab22f03ac74ae7a67dec019c793f11a234ed53e